requirements.txt
asgiref==3.8.1
certifi==2022.12.7
charset-normalizer==3.0.1
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.0.6
django-autoslug==1.9.9
django-autoslug-field==0.2.3
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-email==0.1.10
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
idna==3.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.14
whitenoise==6.7.0
